fix: styles ширина карточек, высота пустого конейнера, надпись если нет товаров,features: Функционал кнопки Показать еще
