+++ src/components/CatalogPage/CatalogPage.css
@@ -83,7 +83,6 @@
   padding: 9px;
   cursor: pointer;
   border: 0.5px solid transparent;
-  height: 30px;
   border-radius: 10px;
 }
 
@@ -96,7 +95,7 @@
   opacity: 0.6;
   cursor: auto;
 }
-
+/* 
 .catalog__pagination {
   align-self: center;
-}
+} */
