+++ src/config/texts.js
@@ -7,3 +7,5 @@ export const CONTACTS_TEXT = 'Контакты';
 export const PHONE = '+7 925 343 34 34';
 export const OFFER_TEXT = 'Оферта';
 export const TERMS_OF_USE_TEXT = 'Политика конфиденциальности';
+
+export const NODATA = 'В данной категории нет товаров';
